web/ttg_backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

django-model-utils==2.4
django==1.9.5
djangorestframework-yaml==1.0.3
djangorestframework==3.3.3
PyYAML==3.11              # via djangorestframework-yaml
